Waypoint Navigation. Use "Settings>Navigation>Waypoint>Load" to load a waypoint file. An example file is provided in src/data/MOEHRING.WPT. Then use the following widgets: WIDGET_WAYPOINT_ARROW, WIDGET_WAYPOINT_DISTANCE, WIDGET_WAYPOINT_TIME to show the direction, distance and time to the next waypoint. Using widget WIDGET_WAYPOINT_INFO shows you the current number and maximum number of waypoints loaded and the name of the next waypoint.
Whenever the pilot reaches the next waypoint, skydrop will go to the next
waypoint, until all waypoints are reached.

Key "<" and ">" on WIDGET_WAYPOINT_INFO can be used to manually go to the
next or previous waypoint.

Please apply patch "smaller-fixes-1" before applying this; as it depends on
that smaller fixes.