Count, store and show distance of a flight (odometer).
The odometer at the beginning of the flight is saved and used to
compute the distance travelled on the flight after the landing.
This data is stored in the log file and shown in the flight log
after selecting a log file during file_details.

This fixes https://github.com/fhorinek/SkyDrop/issues/357

It increases the following sizes:
    gcc version 7.2.0 (Fedora 7.2.0-1.fc28)

    Mem     size-before     size-after      delta    limit
    Flash   186514          186784          +270    196608
    RAM      11922           11930            +8     13107
    EEPROM     773             773             0      2048