Final version for CTR airspace handling
It tells the pilot how far a forbidden CTR is away and in what
direction and height. This allows pilots flying near airports to
ensure, not to enter the forbidden CTR airspace.

It reads airspace data from "/AIR/N48E008.air" (depending on GPS)
and offers 3 new widgets:

 * CTR Arrow:    An arrow pointing away from forbidden CTR. Pilot should
                 always follow that arrow.
 * CTR Distance: The distance to the closed point of the CTR. This
                 tells the pilot how far away he is from CTR.
                 If he is already inside of CTR, then distance to get out.
 * CTR Height:   The height where the CTR starts above the pilot.
                 This allows the pilot to fly below a higher CTR.

If pilot is inside the CTR, the widgets will flash and tell the pilot
to get outside the CTR.

The needed AIR files are precomputed based on data from openaip.net.
They contain for a 64m x 64m raster the needed information
(angle, distance, height), which is used by SkyDrop to only show the
above information.

This code uses very few resources on the SkyDrop, because it only shows
precomputed data. Using gcc version 7.2.0 (Fedora 7.2.0-1.fc28):

    Mem     size-before  size    incr    limit
    Flash   186514       188604  +2090   196608
    RAM     11922        12027   +105    13107
    EEPROM  773          773     +0      2048

The code could be further optimized:
 * combine AGL and airspace handling into combined functions, because
   they are very similar. This would probably need only 1700 bytes of
   flash and 80 bytes of RAM instead of the above values
 * By appending AGL and AIR files into a single file (concat), we will
   only need 1 FIL object, which saves another 30 bytes of RAM.