Reduce memory for sin_table and move to PROGMEM
disp.sin_table is changed to a more compact form, using uint8_t instead
of float to store the values. Additionally it is moved from RAM to PROGMEM
to further save memory.

gcc version 7.2.0 (Fedora 7.2.0-1.fc28)

    Mem     size-before     size-after      delta    limit
    Flash   187342          187290           -52    196608
    RAM      10710           10346          -364     13107
    EEPROM     773             773             0      2048

By using uint8_t instead of float introduces a maximum error
of 0.003802. As "get_sin" is only used to draw pixels on screen,
this minor error is not important.