Fix temperature readings from SH21
I have checked the SHT21 sensor's datasheet and in your code there is a -5 extra value in the temperature calculation function. Like this the temperature output of the vario is inaccurate and shows with 5 Celsius lower temperatures than in real. I have experienced the problem with my vario with the latest firmware. Please fix it. Thanks!